Add comments and reorganize
