+++ lib/msal-browser/src/interaction_client/WamInteractionClient.ts
@@ -13,10 +13,9 @@ import { PopupRequest } from "../request/PopupRequest";
 import { SilentRequest } from "../request/SilentRequest";
 import { SsoSilentRequest } from "../request/SsoSilentRequest";
 import { WamMessageHandler } from "../broker/wam/WamMessageHandler";
-import { WamRequest } from "../request/WamRequest";
 import { WamExtensionMethod } from "../utils/BrowserConstants";
-import { WamExtensionRequestBody } from "../broker/wam/WamExtensionRequest";
-import { WamResponse } from "../response/WamResponse";
+import { WamExtensionRequestBody, WamTokenRequest } from "../broker/wam/WamRequest";
+import { WamResponse } from "../broker/wam/WamResponse";
 import { WamAuthError } from "../error/WamAuthError";
 
 export class WamInteractionClient extends BaseInteractionClient {
@@ -27,6 +26,10 @@ export class WamInteractionClient extends BaseInteractionClient {
         this.provider = provider;
     }
 
+    /**
+     * Acquire token from WAM via browser extension
+     * @param request 
+     */
     async acquireToken(request: PopupRequest|SilentRequest|SsoSilentRequest): Promise<AuthenticationResult> {
         this.logger.trace("WamInteractionClient - acquireToken called.");
         const wamRequest = this.initializeWamRequest(request);
@@ -42,12 +45,22 @@ export class WamInteractionClient extends BaseInteractionClient {
         return this.handleWamResponse(response as WamResponse, wamRequest, reqTimestamp);
     }
 
+    /**
+     * Logout from WAM via browser extension
+     * @param request 
+     */
     logout(request?: EndSessionRequest): Promise<void> {
         this.logger.trace("WamInteractionClient - logout called.");
         return Promise.reject("Logout not implemented yet");
     }
 
-    protected async handleWamResponse(response: WamResponse, request: WamRequest, reqTimestamp: number): Promise<AuthenticationResult> {
+    /**
+     * Transform response from WAM into AuthenticationResult object which will be returned to the end user
+     * @param response 
+     * @param request 
+     * @param reqTimestamp 
+     */
+    protected async handleWamResponse(response: WamResponse, request: WamTokenRequest, reqTimestamp: number): Promise<AuthenticationResult> {
         this.logger.trace("WamInteractionClient - handleWamResponse called.");
         // create an idToken object (not entity)
         const idTokenObj = new AuthToken(response.id_token || Constants.EMPTY_STRING, this.browserCrypto);
@@ -105,7 +118,7 @@ export class WamInteractionClient extends BaseInteractionClient {
      * Translates developer provided request object into WamRequest object
      * @param request 
      */
-    protected initializeWamRequest(request: PopupRequest|SsoSilentRequest): WamRequest {
+    protected initializeWamRequest(request: PopupRequest|SsoSilentRequest): WamTokenRequest {
         this.logger.trace("WamInteractionClient - initializeWamRequest called");
 
         const authority = request.authority || this.config.auth.authority;
@@ -116,7 +129,7 @@ export class WamInteractionClient extends BaseInteractionClient {
         const scopeSet = new ScopeSet(scopes);
         scopeSet.appendScopes(OIDC_DEFAULT_SCOPES);
 
-        const validatedRequest: WamRequest = {
+        const validatedRequest: WamTokenRequest = {
             ...request,
             clientId: this.config.auth.clientId,
             authority: canonicalAuthority.urlString,
